move ear test
