+++ internal/testutil/acc/encryption_at_rest.go
@@ -1,9 +1,14 @@
 package acc
 
 import (
+	"context"
 	"fmt"
+	"strconv"
 
 	"go.mongodb.org/atlas-sdk/v20240805003/admin"
+
+	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
+	"github.com/hashicorp/terraform-plugin-testing/terraform"
 )
 
 func ConfigEARAzureKeyVault(projectID string, azure *admin.AzureKeyVault, useRequirePrivateNetworking, useDatasource bool) string {
@@ -42,4 +47,68 @@ func TestAccDatasourceConfig() string {
 	return `data "mongodbatlas_encryption_at_rest" "test" {
 			project_id = mongodbatlas_encryption_at_rest.test.project_id
 		}`
+}
+
+func TestAccCheckMongoDBAtlasEncryptionAtRestExists(resourceName string) resource.TestCheckFunc {
+	return func(s *terraform.State) error {
+		rs, ok := s.RootModule().Resources[resourceName]
+		if !ok {
+			return fmt.Errorf("not found: %s", resourceName)
+		}
+		if rs.Primary.ID == "" {
+			return fmt.Errorf("no ID is set")
+		}
+		if _, _, err := ConnV2().EncryptionAtRestUsingCustomerKeyManagementApi.GetEncryptionAtRest(context.Background(), rs.Primary.ID).Execute(); err == nil {
+			return nil
+		}
+		return fmt.Errorf("encryptionAtRest (%s) does not exist", rs.Primary.ID)
+	}
+}
+
+func ConvertToAzureKeyVaultEncryptionAtRestAttrMap(az *admin.AzureKeyVault) map[string]string {
+	return map[string]string{
+		"enabled":                    strconv.FormatBool(az.GetEnabled()),
+		"azure_environment":          az.GetAzureEnvironment(),
+		"resource_group_name":        az.GetResourceGroupName(),
+		"key_vault_name":             az.GetKeyVaultName(),
+		"client_id":                  az.GetClientID(),
+		"key_identifier":             az.GetKeyIdentifier(),
+		"subscription_id":            az.GetSubscriptionID(),
+		"tenant_id":                  az.GetTenantID(),
+		"require_private_networking": strconv.FormatBool(az.GetRequirePrivateNetworking()),
+	}
+}
+
+func TestEncryptionAtRestCheckResourceAttr(resourceName, prefix string, attrsMap map[string]string) resource.TestCheckFunc {
+	checks := AddAttrChecksPrefix(resourceName, []resource.TestCheckFunc{}, attrsMap, prefix)
+
+	return resource.ComposeAggregateTestCheckFunc(checks...)
+}
+
+func TestAccCheckMongoDBAtlasEncryptionAtRestDestroy(s *terraform.State) error {
+	for _, rs := range s.RootModule().Resources {
+		if rs.Type != "mongodbatlas_encryption_at_rest" {
+			continue
+		}
+		res, _, err := ConnV2().EncryptionAtRestUsingCustomerKeyManagementApi.GetEncryptionAtRest(context.Background(), rs.Primary.ID).Execute()
+		if err != nil ||
+			(*res.AwsKms.Enabled != false &&
+				*res.AzureKeyVault.Enabled != false &&
+				*res.GoogleCloudKms.Enabled != false) {
+			return fmt.Errorf("encryptionAtRest (%s) still exists: err: %s", rs.Primary.ID, err)
+		}
+	}
+
+	return nil
+}
+
+func TestAccCheckMongoDBAtlasEncryptionAtRestImportStateIDFunc(resourceName string) resource.ImportStateIdFunc {
+	return func(s *terraform.State) (string, error) {
+		rs, ok := s.RootModule().Resources[resourceName]
+		if !ok {
+			return "", fmt.Errorf("not found: %s", resourceName)
+		}
+
+		return rs.Primary.ID, nil
+	}
 }
